Extract editor related code
- in public/js/lib/editor/index.js
